* Added license headers * Refactored re-usable code * added "remove" icon to config list * Made pageTotal count dynamic when downloading services
