+++ projects/ziti-console-lib/src/lib/features/dynamic-widgets/forwarding-config/forwarding-config.service.ts
@@ -15,11 +15,14 @@
 */
 
 import {Injectable} from "@angular/core";
+import {SchemaService} from "../../../services/schema.service";
 
 @Injectable({
     providedIn: 'root'
 })
 export class ForwardingConfigService {
+
+    constructor(private schemaSvc: SchemaService) {}
 
     getProperties(protocol, address, port, forwardProtocol, forwardAddress, forwardPort, allowedProtocols, allowedAddresses, allowedPortRanges) {
         const props = [
@@ -30,80 +33,14 @@ export class ForwardingConfigService {
             {key: 'forwardAddress', value: forwardAddress ? forwardAddress : undefined},
             {key: 'forwardPort', value: forwardPort ? forwardPort : undefined},
             {key: 'allowedAddresses', value: allowedAddresses},
-            {key: 'allowedPortRanges', value: this.getPortRanges(allowedPortRanges)},
+            {key: 'allowedPortRanges', value: this.schemaSvc.getPortRanges(allowedPortRanges)},
             {key: 'allowedProtocols', value: allowedProtocols},
         ];
         return props;
     }
 
-    getPortRanges(allowedPortRanges) {
-        if (!allowedPortRanges) {
-            return [];
-        }
-        const ranges = [];
-        allowedPortRanges.forEach((val: string) => {
-            const vals = val.split('-');
-            if (vals.length === 1) {
-                const port = parseInt(val);
-                ranges.push({high: port, low: port});
-            } else if (vals.length === 2) {
-                const port1 = parseInt(vals[0]);
-                const port2 = parseInt(vals[1]);
-                ranges.push({low: port1, high: port2});
-            } else {
-                // do nothing, invalid range
-            }
-        });
-        return ranges;
-    }
-
-    parseAllowedPortRanges(allowedPortRanges) {
-        const val = [];
-        if (!allowedPortRanges) {
-            return val;
-        }
-        allowedPortRanges.forEach((range: any) => {
-            if (range.low === range.high) {
-                val.push(range.low + '');
-            } else {
-                val.push(range.low + '-' + range.high);
-            }
-        });
-        return val;
-    }
-
     validate(allowedPortRanges) {
         const errors: any = {};
-        errors.allowedPortRanges = this.validatePortRanges(allowedPortRanges);
-    }
-
-    validatePortRanges(allowedPortRanges) {
-        let invalid = false;
-        if (!allowedPortRanges) {
-            return invalid;
-        }
-        allowedPortRanges.forEach((val: string) => {
-            const vals = val.split('-');
-            if (vals.length === 1) {
-                const port = parseInt(val);
-                if (isNaN(port)) {
-                    invalid = true;
-                    return;
-                }
-            } else if (vals.length === 2) {
-                const port1 = parseInt(vals[0]);
-                const port2 = parseInt(vals[1]);
-                if (isNaN(port1) || isNaN(port2)) {
-                    invalid = true;
-                    return;
-                } else if (port1 > port2) {
-                    invalid = true;
-                    return;
-                }
-            } else {
-                // do nothing, invalid range
-            }
-        });
-        return invalid;
+        errors.allowedPortRanges = this.schemaSvc.validatePortRanges(allowedPortRanges);
     }
 }
