+++ projects/ziti-console-lib/src/lib/features/dynamic-widgets/port-ranges/port-ranges.component.ts
@@ -1,5 +1,22 @@
+/*
+    Copyright NetFoundry Inc.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+    https://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+
 import {Component, EventEmitter, Input, Output} from '@angular/core';
 import {isNumber} from 'lodash';
+import {SchemaService} from "../../../services/schema.service";
 
 @Component({
   selector: 'lib-port-ranges',
@@ -13,35 +30,13 @@ export class PortRangesComponent {
   errors: any = {};
   invalid: boolean = false;
 
-  onKeyDown(event?) {
-  }
-
-  get ranges(): any {
-    return {};
-  }
-
-  set ranges(ranges: any) {
-    const portRanges = ranges;
-  }
+  constructor(private schemaSvc: SchemaService) {}
 
   getProperties() {
     if (!this.fieldValue) {
       return [];
     }
-    const ranges = [];
-    this.fieldValue.forEach((val: string) => {
-      const vals = val.split('-');
-      if (vals.length === 1) {
-        const port = parseInt(val);
-        ranges.push({high: port, low: port});
-      } else if (vals.length === 2) {
-        const port1 = parseInt(vals[0]);
-        const port2 = parseInt(vals[1]);
-        ranges.push({low: port1, high: port2});
-      } else {
-        // do nothing, invalid range
-      }
-    });
+    const ranges = this.schemaSvc.getPortRanges(this.fieldValue);
     return [{key: 'portRanges', value: ranges}];
   }
 
@@ -50,15 +45,18 @@ export class PortRangesComponent {
       this.fieldValue = [];
       return;
     }
-    const val = [];
-    ranges.forEach((range: any) => {
-      if (range.low === range.high) {
-        val.push(range.low + '');
-      } else {
-        val.push(range.low + '-' + range.high);
-      }
-    });
-    this.fieldValue = val;
+
+    this.fieldValue = this.schemaSvc.parseAllowedPortRanges(ranges);
+  }
+
+  onKeyup(event: any) {
+    if (event.key === " ") {
+      event.preventDefault();
+      const element = event.target as HTMLElement;
+      element.blur();
+      element.focus();
+    }
+    this.emitEvents();
   }
 
   validateConfig() {
@@ -86,7 +84,7 @@ export class PortRangesComponent {
           return;
         }
       } else {
-        // do nothing, invalid range
+        invalid = true;
       }
     });
     this.invalid = invalid;
