+++ projects/ziti-console-lib/src/lib/features/preview-list/preview-list.component.html
@@ -4,6 +4,7 @@
 
     <div class="text row">
         <div (click)="selected(name)" *ngFor="let name of names" [ngClass]="{ clickable: clickable }" class="listText">
+            <div *ngIf="allowRemove" class="icon-clear" (click)="remove(name)"></div>
             {{ name }}
         </div>
     </div>
