+++ projects/ziti-console-lib/src/lib/features/projectable-forms/service/service-form.component.ts
@@ -1,3 +1,19 @@
+/*
+    Copyright NetFoundry Inc.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+    https://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+
 import {
   Component,
   EventEmitter,
@@ -17,14 +33,13 @@ import {Subscription} from 'rxjs';
 import {ProjectableForm} from "../projectable-form.class";
 import {SETTINGS_SERVICE, SettingsService} from "../../../services/settings.service";
 
-import {isEmpty, forEach, delay, unset, keys, defer, cloneDeep, isEqual, some} from 'lodash';
+import {isEmpty, forEach, delay, unset, keys, defer, cloneDeep, isEqual, some, filter} from 'lodash';
 import {ZITI_DATA_SERVICE, ZitiDataService} from "../../../services/ziti-data.service";
 import {GrowlerService} from "../../messaging/growler.service";
 import {GrowlerModel} from "../../messaging/growler.model";
 import {SERVICE_EXTENSION_SERVICE, ServiceFormService} from './service-form.service';
 import {MatDialogRef} from "@angular/material/dialog";
 import {ExtensionService} from "../../extendable/extensions-noop.service";
-import {SchemaService} from "../../../services/schema.service";
 
 @Component({
   selector: 'lib-service-form',
@@ -38,22 +53,20 @@ import {SchemaService} from "../../../services/schema.service";
   ]
 })
 export class ServiceFormComponent extends ProjectableForm implements OnInit, OnChanges, OnDestroy, AfterViewInit {
-  @Input() formData: any = {};
-  @Input() serviceRoleAttributes: any[] = [];
-  @Output() close: EventEmitter<any> = new EventEmitter<any>();
-  @Output() dataChange: EventEmitter<any> = new EventEmitter<any>();
-
-  paging = {
-    filter: "",
-    noSearch: true,
-    order: "asc",
-    page: 1,
-    searchOn: "name",
-    sort: "name",
-    total: 100
+  @Input() set formData(data) {
+    if (!data?.configs) {
+      data.configs = [];
+    }
+    this.svc.formData = data;
   }
 
-  initData: any = {};
+  get formData(): any {
+    return this.svc.formData;
+  }
+
+  @Input() serviceRoleAttributes: any[] = [];
+  @Output() close: EventEmitter<any> = new EventEmitter<any>();
+
   isEditing = false;
   enrollmentExpiration: any;
   jwt: any;
@@ -74,51 +87,12 @@ export class ServiceFormComponent extends ProjectableForm implements OnInit, OnC
     {id: 'udp', name: 'UDP'},
     {id: 'tcp', name: 'TCP'}
   ];
-  associatedIdentities: any = [];
-  associatedIdentityNames: any = [];
-  identitiesLoading = false;
-  configTypes: any = [];
-  configs: any = [];
-  routers: any = [];
-  filteredConfigs: any = [];
-  selectedConfigTypeId: any = '';
-  selectedConfigType: any = {};
-  selectedSchema: any = {};
-  selectedConfigId: any = '';
-  selectedRouterId: string = '';
-  selectedRouter: string = '';
-  selectedBindingId: string = '';
-  configData: any;
-  addedConfigNames: any = [];
-  addedTerminators: any = [];
-  newConfigName: string = '';
-  terminatorHost: string = '';
-  terminatorPort: string = '';
+
   showMore = false;
-  errors: any = {};
-  configErrors: any = {};
-  terminatorErrors: any = {};
   formView = 'simple';
   settings: any = {};
-  terminatorProtocol = 'udp';
+  errors: any = {};
   subscription: Subscription = new Subscription();
-  hideConfigJSON = false;
-  configJsonView = false;
-  configDataLabel = 'Configuration Form';
-  attachLabel = 'Create and Attach';
-  items: any = [];
-
-  lColorArray = [
-    'black',
-    'white',
-    'black',
-  ]
-
-  bColorArray = [
-    '#33aaff',
-    'var(--secondary)',
-    '#fafafa',
-  ]
 
   @ViewChild("dynamicform", {read: ViewContainerRef}) dynamicForm!: ViewContainerRef;
   constructor(
@@ -126,8 +100,7 @@ export class ServiceFormComponent extends ProjectableForm implements OnInit, OnC
       public svc: ServiceFormService,
       @Inject(ZITI_DATA_SERVICE) private zitiService: ZitiDataService,
       private growlerService: GrowlerService,
-      @Inject(SERVICE_EXTENSION_SERVICE) private extService: ExtensionService,
-      private schemaSvc: SchemaService
+      @Inject(SERVICE_EXTENSION_SERVICE) private extService: ExtensionService
   ) {
     super();
   }
@@ -141,16 +114,14 @@ export class ServiceFormComponent extends ProjectableForm implements OnInit, OnC
     this.jwt = this.formData.enrollmentJwt;
     this.token = this.formData.enrollmentToken;
     this.enrollmentExpiration = this.formData?.enrollmentExpiresAt;
-    this.getAssociatedIdentities();
+    this.svc.getAssociatedConfigs();
+    this.svc.getAssociatedTerminators();
     this.initData = cloneDeep(this.formData);
-    this.watchData();
-    this.extService.updateFormData(this.formData);
     this.subscription.add(
       this.extService.formDataChanged.subscribe((data) => {
         if (data.isEmpty) {
           return;
         }
-        //this.formData = data;
       })
     );
   }
@@ -164,322 +135,27 @@ export class ServiceFormComponent extends ProjectableForm implements OnInit, OnC
     super.ngAfterViewInit();
     this.nameFieldInput.nativeElement.focus();
     this.resetTags();
-    this.getConfigTypes();
-    this.getConfigs().then(() => {
-      this.updatedAddedConfigs();
+    this.svc.getConfigTypes();
+    this.svc.getConfigs().then(() => {
+      this.svc.updatedAddedConfigs();
     });
-    this.getRouters();
+    this.svc.getRouters();
   }
 
   ngOnChanges(changes: SimpleChanges) {
     this.isEditing = !isEmpty(this.formData.id);
   }
 
-  getConfigTypes() {
-    this.zitiService.get('config-types', this.paging, []).then((result: any) => {
-      this.configTypes = result.data;
-    });
+  configChanged($event) {
+    this.svc.configChanged(this.dynamicForm);
   }
 
-  getConfigs() {
-    return this.zitiService.get('configs', this.paging, []).then((result: any) => {
-      this.configs = result.data;
-      this.configTypeChanged();
-    });
-  }
-
-  getRouters() {
-    this.zitiService.get('edge-routers', this.paging, []).then((result: any) => {
-      this.routers = result.data;
-    });
-  }
-
-  updatedAddedConfigs() {
-    this.addedConfigNames = [];
-    this.configs.forEach((availableConfig) => {
-      const cfgExists = some(this.formData.configs, configId => {
-        return availableConfig.id == configId;
-      });
-      if (cfgExists) {
-        this.addedConfigNames.push(availableConfig.name);
-        this.addedConfigNames = [...this.addedConfigNames];
-      }
-    })
-  }
-
-  toggleJSONView() {
-    this.configJsonView = !this.configJsonView;
-    this.configDataLabel = this.configJsonView ? 'JSON Configuration' : 'Configuration Form';
-    this.updateConfigData();
-  }
-
-  updateConfigData() {
-    if (!this.configJsonView) {
-      this.updateFormView(this.items, this.configData);
-    } else {
-      this.hideConfigJSON = true;
-      defer(() => {
-        const data = {};
-        this.addItemsToConfig(this.items, data);
-        this.configData = data;
-        this.hideConfigJSON = false;
-      });
-    }
-  }
-
-  updateFormView(items, data) {
-    items.forEach((item) => {
-      if (item.items) {
-        this.updateFormView(item.items, data[item.key]);
-      } else if (item?.component?.instance?.setProperties) {
-        let val;
-        switch (item.key) {
-          case 'forwardingconfig':
-            val = {
-              protocol: data.protocol,
-              address: data.address,
-              port: data.port,
-              forwardProtocol: data.forwardProtocol,
-              forwardAddress: data.forwardAddress,
-              forwardPort: data.forwardPort,
-              allowedProtocols: data.allowedProtocols,
-              allowedAddresses: data.allowedAddresses,
-              allowedPortRanges: data.allowedPortRanges
-            }
-            break;
-          default:
-            val = data[item.key];
-            break;
-        }
-        item?.component?.instance?.setProperties(val);
-      } else if (item?.component?.setInput) {
-        item.component.setInput('fieldValue', data[item.key]);
-      }
-    });
-    return data;
-  }
-
-  addItemsToConfig(items, data) {
-    items.forEach((item) => {
-      let props = [];
-      if (item.items) {
-        data[item.key] = this.addItemsToConfig(item.items, {});
-      } else if (item?.component?.instance?.getProperties) {
-        props = item?.component?.instance?.getProperties();
-      } else if (item?.component?.instance) {
-        props = [{key: item.key, value: item.component.instance.fieldValue}];
-      }
-      props.forEach((prop) => {
-        data[prop.key] = prop.value;
-      });
-    });
-    return data;
-  }
-
-  configTypeChanged(event?: any) {
-    this.filteredConfigs = this.configs.filter((config) => {
-      return config.configTypeId === this.selectedConfigTypeId;
-    });
-    this.configTypes.forEach((configType) => {
-      if (this.selectedConfigTypeId === configType.id) {
-        this.selectedConfigType = configType;
-      }
-    });
-    this.selectedConfigId = !isEmpty(this.selectedConfigTypeId) ? 'add-new' : '';
-    this.configChanged();
-  }
-
-  async configChanged(event?: any) {
-    let selectedConfig: any = {};
-    this.configData = undefined;
-    let data;
-    let attachLabel = 'Attach to Service';
-    if (this.selectedConfigId === 'add-new') {
-      data = {};
-      this.selectedSchema = await this.zitiService.schema(this.selectedConfigType.schema);
-      await this.createForm();
-      attachLabel = 'Create and Attach';
-    } else if (this.selectedConfigId) {
-      this.filteredConfigs.forEach((config) => {
-        if (this.selectedConfigId === config.id) {
-          selectedConfig = config;
-        }
-      });
-      data = selectedConfig?.data || {};
-    }
-    if (!this.configData) {
-      this.configData = data;
-    } else {
-      defer(() => {
-        this.configData = cloneDeep(data);
-      });
-    }
-    this.updateConfigData();
-    this.attachLabel = attachLabel;
-  }
-
-  async createForm() {
-    this.clearForm();
-    if (this.selectedConfigType && this.dynamicForm) {
-      if (this.selectedSchema) {
-        this.renderSchmea(this.selectedSchema);
-      }
-    }
-  }
-
-  renderSchmea(schema: any) {
-    if (schema.properties) {
-      this.items = this.schemaSvc.render(schema, this.dynamicForm, this.lColorArray, this.bColorArray);
-      for (let obj of this.items) {
-        const cRef = obj.component;
-        cRef.instance.errors = this.errors;
-        if (cRef?.instance.valueChange) {
-          const pName: string[]  = cRef.instance.parentage;
-          let parentKey;
-          if(pName) parentKey = pName.join('.');
-          if (parentKey && !this.formData[parentKey]) this.formData[parentKey] = {};
-        }
-      }
-    }
-  }
-
-  clearForm() {
-    this.items.forEach((item: any) => {
-      if (item?.component) item.component.destroy();
-    });
-    this.errors = {};
-    this.items = [];
-    this.formData = {
-      configs: [],
-      encryptionRequired: true,
-      terminatorStrategy: '',
-    };
-    if (this.subscription) this.subscription.unsubscribe();
-  }
-
-  async attachConfig(addedConfigId) {
-    if (this.selectedConfigId === 'add-new') {
-      if (!this.validateConfig()) {
-        return;
-      }
-      const newConfig = {
-        configTypeId: this.selectedConfigTypeId,
-        data: this.configData,
-        name: this.newConfigName
-      }
-      const configId = await this.svc.createConfig(newConfig)
-          .then((result) => {
-            return result?.data?.id;
-          })
-          .catch((result) => {
-        const errorField = result?.error?.error?.cause?.field;
-        if (!isEmpty(errorField)) {
-          this.configErrors[errorField] = true;
-        }
-        const errorMessage = result?.error?.error?.cause?.reason;
-        const growlerData = new GrowlerModel(
-            'error',
-            'Error',
-            `Error Creating Config`,
-            errorMessage,
-        );
-        this.growlerService.show(growlerData);
-        return undefined;
-      });
-      if (!isEmpty(configId)) {
-        this.formData.configs.push(configId);
-        this.addedConfigNames.push(this.newConfigName);
-        this.addedConfigNames = [...this.addedConfigNames];
-        this.getConfigs();
-        const growlerData = new GrowlerModel(
-            'success',
-            'Success',
-            `New Config Attached`,
-            `New Config ${this.newConfigName} has been created and attached to the service`,
-        );
-        this.growlerService.show(growlerData);
-        this.newConfigName = '';
-        this.selectedConfigTypeId = '';
-        this.selectedConfigId = '';
-        return;
-      }
-    }
-    let configAdded = false;
-    this.formData.configs.forEach((configId) => {
-      if (configId === addedConfigId) {
-        configAdded = true;
-      }
-    });
-    if (!configAdded) {
-      let configName = '';
-      this.configs.forEach((config) => {
-        if (config.id === addedConfigId) {
-          configName = config.name;
-        }
-      });
-      if (!isEmpty(configName)) {
-        this.addedConfigNames.push(configName);
-        this.addedConfigNames = [...this.addedConfigNames];
-      }
-      this.formData.configs.push(addedConfigId);
-    } else {
-      const growlerData = new GrowlerModel(
-          'warning',
-          'Info',
-          `Config Already Attached`,
-          'Config has already been attached to this service',
-      );
-      this.growlerService.show(growlerData);
-    }
-  }
-
-  routerChanged(event?: any) {
-    let selectedRouter;
-    if (this.selectedConfigId) {
-      this.routers.forEach((router) => {
-        if (this.selectedRouterId === router.id) {
-          selectedRouter = router;
-        }
-      });
-    }
-    this.selectedRouter = selectedRouter;
-  }
-
-  addTerminator(item) {
-    let termAdded = false;
-    this.addedTerminators.forEach((termName) => {
-      if (item.name === termName) {
-        termAdded = true;
-      }
-    });
-    if (!termAdded) {
-      this.addedTerminators.push(item.name);
-    }
-  }
-
-  bindingTypeChanged(event?: any) {
-
-  }
-
-  terminatorProtocolChanged(event?: any) {
-
+  configTypeChanged($event) {
+    this.svc.configTypeChanged(this.dynamicForm);
   }
 
   get showConfigData() {
-    return this.selectedConfigId === 'add-new';
-  }
-
-  get showAttachButton() {
-    return !isEmpty(this.selectedConfigId);
-  }
-
-  getAssociatedIdentities() {
-    this.zitiService.getSubdata('services', this.formData.id, 'identities').then((result: any) => {
-      this.associatedIdentities = result.data;
-      this.associatedIdentityNames = this.associatedIdentities.map((svc) => {
-        return svc.name;
-      });
-    });
+    return this.svc.selectedConfigId === 'add-new';
   }
 
   headerActionRequested(action) {
@@ -497,65 +173,35 @@ export class ServiceFormComponent extends ProjectableForm implements OnInit, OnC
   }
 
   async save(event?) {
-    const isValid = this.validate();
+    const isValid = this.svc.validate();
     const isExtValid = await this.extService.validateData();
+    const isEdit = !isEmpty(this.formData.id);
     if(!isValid || !isExtValid) {
       return;
     }
 
     this.isLoading = true;
-    this.svc.save(this.formData).then((result) => {
-      if (!isEmpty(result.id)) {
+    const serviceId = await this.svc.save(this.formData).then((result) => {
+      if (!isEmpty(result?.id)) {
         this.formData = result;
         this.initData = this.formData;
       }
+      return result?.id;
     }).finally(() => {
       this.isLoading = false;
     });
-  }
-
-  validate() {
-    this.errors = {};
-    if (isEmpty(this.formData.name)) {
-      this.errors['name'] = true;
+    if (serviceId && !isEdit) {
+      await this.svc.addTerminators(serviceId)
     }
-    return isEmpty(this.errors);
-  }
-
-  validateConfig() {
-    this.configErrors = {};
-    if (isEmpty(this.newConfigName)) {
-      this.configErrors['name'] = true;
+    if (serviceId) {
+      this.closeModal(true, true);
     }
-    return isEmpty(this.configErrors);
   }
 
   toggleEncryptionRequired() {
     this.formData.encryptionRequired = !this.formData.encryptionRequired;
   }
 
-  closeModal(refresh = true, ignoreChanges = false): void {
-    if (!ignoreChanges && this._dataChange) {
-      const confirmed = confirm('You have unsaved changes. Do you want to leave this page and discard your changes or stay on this page?');
-      if (!confirmed) {
-        return;
-      }
-    }
-    this.close.emit({refresh: refresh});
-  }
-
   clear(): void {
-  }
-
-  _dataChange = false;
-  watchData() {
-    delay(() => {
-      const dataChange = !isEqual(this.initData, this.formData);
-      if (dataChange !== this._dataChange) {
-        this.dataChange.emit(dataChange);
-      }
-      this._dataChange = dataChange;
-      this.watchData();
-    }, 100);
   }
 }
