+++ projects/ziti-console-lib/src/lib/models/service.ts
@@ -2,6 +2,7 @@ export class Service {
     name = '';
     encryptionRequired = true;
     terminatorStrategy = '';
+    roleAttributes: any[] = [];
     tags: any = {};
     configs: any[] = []
 };
