+++ projects/ziti-console-lib/src/lib/pages/services/services-page.component.ts
@@ -1,3 +1,19 @@
+/*
+    Copyright NetFoundry Inc.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+    https://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+
 import {Component, Inject, OnInit, OnDestroy} from '@angular/core';
 import {ServicesPageService} from "../services/services-page.service";
 import {DataTableFilterService} from "../../features/data-table/data-table-filter.service";
