+++ projects/ziti-console-lib/src/lib/services/schema.service.ts
@@ -20,7 +20,7 @@ import {NumberInputComponent} from "../features/dynamic-widgets/number/number-in
 import {BooleanToggleInputComponent} from "../features/dynamic-widgets/boolean/boolean-toggle-input.component";
 import {StringInputComponent} from "../features/dynamic-widgets/string/string-input.component";
 import {SelectorInputComponent} from "../features/dynamic-widgets/selector/selector-input.component";
-import _ from "lodash";
+import _, {isNumber} from "lodash";
 import {TextListInputComponent} from "../features/dynamic-widgets/text-list/text-list-input.component";
 import {CheckboxListInputComponent} from "../features/dynamic-widgets/checkbox-list/checkbox-list-input.component";
 import {
@@ -652,6 +652,76 @@ export class SchemaService {
             componentRef = this.buildTextField(view, nestLevel, key, parentage);
         }
         return componentRef;
+    }
+
+    getPortRanges(allowedPortRanges) {
+        if (!allowedPortRanges) {
+            return [];
+        }
+        const ranges = [];
+        allowedPortRanges.forEach((val: string) => {
+            const vals = val.split('-');
+            if (vals.length === 1) {
+                const port = parseInt(val);
+                ranges.push({high: port, low: port});
+            } else if (vals.length === 2) {
+                const port1 = parseInt(vals[0]);
+                const port2 = parseInt(vals[1]);
+                ranges.push({low: port1, high: port2});
+            } else {
+                // do nothing, invalid range
+            }
+        });
+        return ranges;
+    }
+
+    parseAllowedPortRanges(allowedPortRanges) {
+        const val = [];
+        if (!allowedPortRanges) {
+            return val;
+        }
+        allowedPortRanges.forEach((range: any) => {
+            if (range.low === range.high) {
+                val.push(range.low + '');
+            } else {
+                val.push(range.low + '-' + range.high);
+            }
+        });
+        return val;
+    }
+
+    validatePortRanges(allowedPortRanges) {
+        let invalid = false;
+        if (!allowedPortRanges) {
+            return invalid;
+        }
+        allowedPortRanges.forEach((val: string) => {
+            const vals = val.split('-');
+            if (vals.length === 1) {
+                const port = parseInt(val);
+                if (!this.isValidPort(port)) {
+                    invalid = true;
+                    return;
+                }
+            } else if (vals.length === 2) {
+                const port1 = parseInt(vals[0]);
+                const port2 = parseInt(vals[1]);
+                if (!this.isValidPort(port1) || !this.isValidPort(port2)) {
+                    invalid = true;
+                    return;
+                } else if (port1 > port2) {
+                    invalid = true;
+                    return;
+                }
+            } else {
+                invalid = true;
+            }
+        });
+        return invalid;
+    }
+
+    isValidPort(port) {
+        return !isNaN(port) && isNumber(port) && port >= 1 && port <= 65535;
     }
 
     private addError(errors: any, key: string, msg: string) {
